+++ astropy/utils/__init__.py
@@ -12,5 +12,8 @@ from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
 
+from .codegen import *
 from .compat.odict import OrderedDict
+from .decorators import *
+from .introspection import *
 from .misc import *
