+++ astropy/utils/introspection.py
@@ -1,0 +1,196 @@
+# -*- coding: utf-8 -*-
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""Functions related to Python runtime introspection."""
+
+
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
+
+
+import inspect
+import sys
+
+from ..extern import six
+
+
+__all__ = ['find_current_module']
+
+
+__doctest_skip__ = ['find_current_module']
+
+
+def find_current_module(depth=1, finddiff=False):
+    """
+    Determines the module/package from which this function is called.
+
+    This function has two modes, determined by the ``finddiff`` option. it
+    will either simply go the requested number of frames up the call
+    stack (if ``finddiff`` is False), or it will go up the call stack until
+    it reaches a module that is *not* in a specified set.
+
+    Parameters
+    ----------
+    depth : int
+        Specifies how far back to go in the call stack (0-indexed, so that
+        passing in 0 gives back `astropy.utils.misc`).
+    finddiff : bool or list
+        If False, the returned ``mod`` will just be ``depth`` frames up from
+        the current frame. Otherwise, the function will start at a frame
+        ``depth`` up from current, and continue up the call stack to the
+        first module that is *different* from those in the provided list.
+        In this case, ``finddiff`` can be a list of modules or modules
+        names. Alternatively, it can be True, which will use the module
+        ``depth`` call stack frames up as the module the returned module
+        most be different from.
+
+    Returns
+    -------
+    mod : module or None
+        The module object or None if the package cannot be found. The name of
+        the module is available as the ``__name__`` attribute of the returned
+        object (if it isn't None).
+
+    Raises
+    ------
+    ValueError
+        If ``finddiff`` is a list with an invalid entry.
+
+    Examples
+    --------
+    The examples below assume that there are two modules in a package named
+    ``pkg``. ``mod1.py``::
+
+        def find1():
+            from astropy.utils import find_current_module
+            print find_current_module(1).__name__
+        def find2():
+            from astropy.utils import find_current_module
+            cmod = find_current_module(2)
+            if cmod is None:
+                print 'None'
+            else:
+                print cmod.__name__
+        def find_diff():
+            from astropy.utils import find_current_module
+            print find_current_module(0,True).__name__
+
+    ``mod2.py``::
+
+        def find():
+            from .mod1 import find2
+            find2()
+
+    With these modules in place, the following occurs::
+
+        >>> from pkg import mod1, mod2
+        >>> from astropy.utils import find_current_module
+        >>> mod1.find1()
+        pkg.mod1
+        >>> mod1.find2()
+        None
+        >>> mod2.find()
+        pkg.mod2
+        >>> find_current_module(0)
+        <module 'astropy.utils.misc' from 'astropy/utils/misc.py'>
+        >>> mod1.find_diff()
+        pkg.mod1
+
+    """
+
+    frm = inspect.currentframe()
+    for i in range(depth):
+        frm = frm.f_back
+        if frm is None:
+            return None
+
+    if finddiff:
+        currmod = inspect.getmodule(frm)
+        if finddiff is True:
+            diffmods = [currmod]
+        else:
+            diffmods = []
+            for fd in finddiff:
+                if inspect.ismodule(fd):
+                    diffmods.append(fd)
+                elif isinstance(fd, six.string_types):
+                    diffmods.append(__import__(fd))
+                elif fd is True:
+                    diffmods.append(currmod)
+                else:
+                    raise ValueError('invalid entry in finddiff')
+
+        while frm:
+            frmb = frm.f_back
+            modb = inspect.getmodule(frmb)
+            if modb not in diffmods:
+                return modb
+            frm = frmb
+    else:
+        return inspect.getmodule(frm)
+
+
+def find_mod_objs(modname, onlylocals=False):
+    """ Returns all the public attributes of a module referenced by name.
+
+    .. note::
+        The returned list *not* include subpackages or modules of
+        ``modname``, nor does it include private attributes (those that
+        beginwith '_' or are not in `__all__`).
+
+    Parameters
+    ----------
+    modname : str
+        The name of the module to search.
+    onlylocals : bool or list of str
+        If `True`, only attributes that are either members of ``modname`` OR
+        one of its modules or subpackages will be included. If it is a list
+        of strings, those specify the possible packages that will be
+        considered "local".
+
+    Returns
+    -------
+    localnames : list of str
+        A list of the names of the attributes as they are named in the
+        module ``modname`` .
+    fqnames : list of str
+        A list of the full qualified names of the attributes (e.g.,
+        ``astropy.utils.introspection.find_mod_objs``). For attributes that are
+        simple variables, this is based on the local name, but for functions or
+        classes it can be different if they are actually defined elsewhere and
+        just referenced in ``modname``.
+    objs : list of objects
+        A list of the actual attributes themselves (in the same order as
+        the other arguments)
+
+    """
+
+    __import__(modname)
+    mod = sys.modules[modname]
+
+    if hasattr(mod, '__all__'):
+        pkgitems = [(k, mod.__dict__[k]) for k in mod.__all__]
+    else:
+        pkgitems = [(k, mod.__dict__[k]) for k in dir(mod) if k[0] != '_']
+
+    # filter out modules and pull the names and objs out
+    ismodule = inspect.ismodule
+    localnames = [k for k, v in pkgitems if not ismodule(v)]
+    objs = [v for k, v in pkgitems if not ismodule(v)]
+
+    # fully qualified names can be determined from the object's module
+    fqnames = []
+    for obj, lnm in zip(objs, localnames):
+        if hasattr(obj, '__module__') and hasattr(obj, '__name__'):
+            fqnames.append(obj.__module__ + '.' + obj.__name__)
+        else:
+            fqnames.append(modname + '.' + lnm)
+
+    if onlylocals:
+        if onlylocals is True:
+            onlylocals = [modname]
+        valids = [any([fqn.startswith(nm) for nm in onlylocals]) for fqn in fqnames]
+        localnames = [e for i, e in enumerate(localnames) if valids[i]]
+        fqnames = [e for i, e in enumerate(fqnames) if valids[i]]
+        objs = [e for i, e in enumerate(objs) if valids[i]]
+
+    return localnames, fqnames, objs
